requirements_minimal.txt
# Requirements for Minimal Installation
# Minimal Requirements for Ransomware Protection System

# Core Web Framework
flask>=2.0.0
flask-cors>=4.0.0
waitress>=2.0.0

# File System Monitoring (CRITICAL - Main Issue)
watchdog>=2.0.0

# System Utilities
psutil>=5.8.0
requests>=2.25.0

# Data Processing
numpy>=1.20.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0

# Configuration
pyyaml>=6.0

# Cryptography
cryptography>=3.4.8

# Security Analysis (CRITICAL - YARA Scanner)
yara-python>=4.3.0

# Logging
colorama>=0.4.6